refactor sources & change namespace scope
